WIP - using add only
