+++ lib/story_branch/commands/add.rb
@@ -9,28 +9,40 @@ module StoryBranch
     class Add < StoryBranch::Command
       def initialize(options)
         @options = options
-        @config = init_config
+        @config = init_config(ENV['HOME'])
       end
 
       def execute(input: $stdin, output: $stdout)
-        append_to_config
+        create_global_config
+        create_local_config
+        output.puts 'Configuration added successfully'
       end
 
       private
 
-      def append_to_config
-        prompt = ::TTY::Prompt.new
-        project_name = prompt.ask "What should be this project's name?"
+      def create_local_config
+        local_config = init_config('.')
+        local_config.set(:project_name, value: project_name)
+        local_config.write
+      end
+
+      def create_global_config
         api_key = prompt.ask 'Please provide the api key:'
         project_id = prompt.ask "Please provide this project's id:"
         @config.set(project_name, :api_key, value: api_key)
         @config.set(project_name, :project_id, value: project_id)
-        @config.write(force: true)
+        @config.write
       end
 
-      # TODO: Move somewhere else as it is common across multiple commands
-      def init_config
-        config_file_path = Dir.home
+      def project_name
+        return @project_name if @project_name
+        prompt = ::TTY::Prompt.new
+        @project_name = prompt.ask "What should be this project's name?"
+        @project_name
+      end
+
+      def init_config(path)
+        config_file_path = path
         config_file_name = '.story_branch'
         config = ::TTY::Config.new
         config.filename = config_file_name
